+++ zarrs/src/array/array_bytes.rs
@@ -7,8 +7,8 @@ use unsafe_cell_slice::UnsafeCellSlice;
 
 use crate::{
     array_subset::{ArraySubset, IncompatibleArraySubsetAndShapeError},
-    byte_range::extract_byte_ranges_concat_unchecked,
     metadata::DataTypeSize,
+    storage::byte_range::extract_byte_ranges_concat_unchecked,
 };
 
 use super::{