+++ zarrs/src/array/codec/bytes_to_bytes/blosc.rs
@@ -269,7 +269,7 @@ mod tests {
             ArrayRepresentation, BytesRepresentation, DataType,
         },
         array_subset::ArraySubset,
-        byte_range::ByteRange,
+        storage::byte_range::ByteRange,
     };
 
     use super::*;