+++ CHANGELOG.md
@@ -90,6 +90,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: Remove `Par{Chunks,Indices}IteratorProducer`, which were unneeded
 - **Breaking**: Remove `[Async]BytesPartialDecoderTraits` implementations for `std::io::Cursor` variants
 - **Breaking**: Remove `ArraySubset::chunks()` and `array_subset::iterators::Chunks`
+- **Breaking**: Remove `storage::byte_range` re-export
 
 ### Fixed
 - Permit data types with empty configurations that do not require one